Make api & SQL compile
