+++ sql/core/src/main/scala/org/apache/spark/sql/package.scala
@@ -17,12 +17,8 @@
 
 package org.apache.spark
 
-import java.util.regex.Pattern
-
 import org.apache.spark.annotation.{DeveloperApi, Unstable}
-import org.apache.spark.sql.catalyst.trees.{CurrentOrigin, Origin, PySparkCurrentOrigin}
 import org.apache.spark.sql.execution.SparkStrategy
-import org.apache.spark.sql.internal.SQLConf
 
 /**
  * Allows the execution of relational queries, including those expressed in SQL using Spark.
@@ -77,47 +73,4 @@ package object sql {
    * with rebasing.
    */
   private[sql] val SPARK_LEGACY_INT96_METADATA_KEY = "org.apache.spark.legacyINT96"
-
-  /**
-   * This helper function captures the Spark API and its call site in the user code from the current
-   * stacktrace.
-   *
-   * As adding `withOrigin` explicitly to all Spark API definition would be a huge change,
-   * `withOrigin` is used only at certain places where all API implementation surely pass through
-   * and the current stacktrace is filtered to the point where first Spark API code is invoked from
-   * the user code.
-   *
-   * As there might be multiple nested `withOrigin` calls (e.g. any Spark API implementations can
-   * invoke other APIs) only the first `withOrigin` is captured because that is closer to the user
-   * code.
-   *
-   * @param f The function that can use the origin.
-   * @return The result of `f`.
-   */
-  private[sql] def withOrigin[T](f: => T): T = {
-    if (CurrentOrigin.get.stackTrace.isDefined) {
-      f
-    } else {
-      val st = Thread.currentThread().getStackTrace
-      var i = 0
-      // Find the beginning of Spark code traces
-      while (i < st.length && !sparkCode(st(i))) i += 1
-      // Stop at the end of the first Spark code traces
-      while (i < st.length && sparkCode(st(i))) i += 1
-      val origin = Origin(stackTrace = Some(st.slice(
-        from = i - 1,
-        until = i + SQLConf.get.stackTracesInDataFrameContext)),
-        pysparkErrorContext = PySparkCurrentOrigin.get())
-      CurrentOrigin.withOrigin(origin)(f)
-    }
-  }
-
-  private val sparkCodePattern = Pattern.compile("(org\\.apache\\.spark\\.sql\\." +
-      "(?:functions|Column|ColumnName|SQLImplicits|Dataset|DataFrameStatFunctions|DatasetHolder)" +
-      "(?:|\\..*|\\$.*))" +
-      "|(scala\\.collection\\..*)")
-
-  private def sparkCode(ste: StackTraceElement): Boolean = {
-    sparkCodePattern.matcher(ste.getClassName).matches()
-  }
 }
